File renaming and JSDoc comments
