+++ pages/about.js
@@ -1,6 +1,6 @@
 import Head from "next/head"
 import styles from "../styles/About.module.sass"
-import Panorama from "../components/Panorama/Panorama"
+import Panorama from "../components/Panorama"
 import Footer from "../components/Footer"
 import ListItem from "../components/About/ListItem"
 import Button from "../components/Button"
@@ -10,7 +10,6 @@ import Button from "../components/Button"
  *
  * @author Eric Keränen
  * @returns {JSX.Element}
- * @constructor
  */
 export default function About() {
     return (
