+++ pages/index.js
@@ -1,6 +1,6 @@
 import Head from "next/head"
 import styles from "../styles/Home.module.sass"
-import Panorama from "../components/Panorama/Panorama"
+import Panorama from "../components/Panorama"
 import Footer from "../components/Footer"
 import Button from "../components/Button"
 
@@ -9,7 +9,6 @@ import Button from "../components/Button"
  *
  * @author Eric Keränen
  * @returns {JSX.Element}
- * @constructor
  */
 export default function Home() {
   return (
